giving up on spidering for now. see note below
This was supposed to be a class that would let you load a yaml file
containing !include and !relpath tags, and copy all those dependencies
to a new folder, and then modify the paths in the yaml file so that it
could still be loaded.

Untested, but I think now it'll load yaml recursively and make a note
of all the !include/!relpath references, but not change anything.
Still need to figure out how to modify the paths in a sane manner i.e.
what folder structure do we copy to? Do we path-mangle for duplicate
file names? What about directory paths, symlinks, etc?
Then I'd have to re-write experiment_tenn2.py to actually use this new
spider loader when loading RSS project files...